Moved `verify_binary_contains_lints` to `tests/`
