+++ Cargo.toml
@@ -37,4 +37,6 @@ bugreport = "0.5.0"
 itertools = "0.10.5"
 cargo_toml = "0.13.0"
 toml = "0.5.9"
+
+[dev-dependencies]
 assert_cmd = "2.0"
